pyproject.toml
[tool.poetry]
name = "bonobrownie"
version = "0.1.0"
description = "Probono project for small business"
authors = ["Luis Maia <luismaiasombra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.118.0"}
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.11.0"
sqlalchemy = "^2.0.43"
requests = "^2.32.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
